[INIT] new project for refactoring
