--- README.md
@@ -1,3 +1,0 @@
-# DSMBamBoo-Backend 
-[](https://img.shields.io/website?down_color=lightgrey&down_message=offline&style=plastic&up_color=blue&up_message=online&url=https%3A%2F%2Fapi.dsmbamboo.com%2Fapi%2Fv1%2Factuator%2Fhealth)
-대덕소프트웨어마이스터고등학교 대나무숲 Backend API
