+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'dsmforest'
+rootProject.name = 'api'
