--- src/main/java/com/dsmbamboo/dsmforest/domain/Notice.java
@@ -1,22 +1,0 @@
-package com.dsmbamboo.dsmforest.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import java.time.LocalDateTime;
-
-@Entity
-@Getter @Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class Notice {
-
-    @Id private long id;
-
-    private String title;
-    private LocalDateTime createdDate;
-}
