+++ src/test/java/com/dsmbamboo/api/ApiApplicationTests.java
@@ -1,0 +1,13 @@
+package com.dsmbamboo.api;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ApiApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
