--- src/test/java/com/dsmbamboo/dsmforest/DsmforestApplicationTests.java
@@ -1,16 +1,0 @@
-package com.dsmbamboo.dsmforest;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class DsmforestApplicationTests {
-
-    @Test
-    public void contextLoads() {
-    }
-
-}
